Implement toolbox fix docs (#266)
* fix: docs

* fix: docs
