+++ fastagency/models/__init__.py
@@ -1,6 +1,6 @@
 # from .registry import Registry  # ModelSchema, ModelSchemas, Registry, Schemas
 
-from . import agents, applications  # noqa: F401
+from . import agents, applications, llms, secrets, teams, toolboxes  # noqa: F401
 from .registry import Registry
 
 __all__ = ["Registry"]
