--- fastagency/models/secrets/__init__.py
@@ -1,6 +1,0 @@
-from .openapi_auth import OpenAPIAuth, OpenAPIAuthRef
-
-__all__ = [
-    "OpenAPIAuth",
-    "OpenAPIAuthRef",
-]
