+++ scripts/build-docs.sh
@@ -3,4 +3,5 @@
 set -e
 set -x
 
+rm -rf docs/docs/en/api
 cd docs; python docs.py build
